backend/requirements-audio.txt
# Audio Processing Requirements
# Simplified for Python 3.14 compatibility

# Core audio processing
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0

# Speech-to-Text - OpenAI Whisper (simpler, no Rust required)
openai-whisper
# Note: For production, consider using OpenAI API instead

# Text-to-Speech (Python 3.14 compatible)
pyttsx3>=2.90  # Offline TTS (Windows SAPI)
gTTS  # Google TTS (requires internet)
edge-tts  # Microsoft Edge TTS (free, high quality)

# Audio format conversion
pydub>=0.25.1
librosa>=0.10.1

# WebRTC for browser audio
aiortc>=1.6.0  # WebRTC implementation
av>=11.0.0  # Audio/video handling

# WebSocket for real-time audio streaming
python-multipart>=0.0.6
